Moving theme into design system module.
