+++ app/src/main/java/com/adammcneilly/pocketleague/ui/EventDetailContent.kt
@@ -46,6 +46,7 @@ import coil.compose.rememberAsyncImagePainter
 import coil.imageLoader
 import coil.request.ImageRequest
 import coil.request.SuccessResult
+import com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme
 import com.adammcneilly.pocketleague.core.displaymodels.EventDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.EventStageSummaryDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.TeamOverviewDisplayModel
@@ -54,7 +55,6 @@ import com.adammcneilly.pocketleague.ui.components.Tooltip
 import com.adammcneilly.pocketleague.ui.composables.components.Chip
 import com.adammcneilly.pocketleague.ui.composables.eventstage.StageSummaryListItem
 import com.adammcneilly.pocketleague.ui.composables.team.TeamOverviewListItem
-import com.adammcneilly.pocketleague.ui.theme.PocketLeagueTheme
 import com.google.accompanist.flowlayout.FlowRow
 import com.google.accompanist.placeholder.material.placeholder
 import kotlinx.coroutines.launch
@@ -381,7 +381,7 @@ private fun EventDetailContentPreview() {
         eventDetail = eventDetail,
     )
 
-    PocketLeagueTheme {
+    com.adammcneilly.pocketleague.android.designsystem.theme.PocketLeagueTheme {
         Surface {
             EventDetailContent(
                 viewState = viewState,
